refactor(expenditures): move to declarative route table via @branch/lambda-http
Replaces the if-chain in handler.ts with dispatch() + a routes.ts table,
matching the shared @branch/lambda-http package adopted repo-wide.

- handler.ts: now just `dispatch(event, { prefix: 'expenditures', routes })`.
- routes.ts: ordered Route[] table, ROUTES-START/END markers preserved
  around the entries. Route order matches the original if-chain, notably
  keeping /expenditures/upload-url before /expenditures/:id.
- controllers/expenditures.ts: one RouteHandler per route — validates
  input, calls the service layer, shapes the response. Same status codes,
  messages, and validation order as before.
- services/expenditures.ts: Kysely queries, S3 presigning, and
  receiptKeyFromUrl, unchanged in behavior, just relocated.
- Local json()/requireAuth() dropped in favor of the @branch/lambda-http
  exports (requireAuth's ADMIN gate on PATCH /expenditures/:id/status is
  now backed by the real @branch/lambda-auth checkAuthorization instead
  of the handler-local wrapper; identical logic).
- package.json: added @branch/lambda-http as a dependency;
  package-lock.json regenerated via `npm install --legacy-peer-deps`.
- test/expenditures.unit.test.ts: added a route-precedence regression
  test for GET /expenditures/upload-url vs /expenditures/:id.

No behavior change: same status codes, response shapes, S3 TTLs, and
content-type restriction as the previous if-chain.
