+++ apps/backend/lambdas/expenditures/package.json
@@ -29,6 +29,7 @@
     "@aws-sdk/client-s3": "^3.995.0",
     "@aws-sdk/s3-request-presigner": "^3.995.0",
     "@branch/lambda-auth": "file:../../../../shared/lambda-auth",
+    "@branch/lambda-http": "file:../../../../shared/lambda-http",
     "aws-jwt-verify": "^5.1.1",
     "aws-lambda": "^1.0.7",
     "kysely": "^0.28.8",
